rtc: rtc-fake: Fix reading time
If the real seconds value is set on boot or resume
the next call of rtc_fake_read_time will re-set it
adding again the initial value rtc_time_sec.
So we'll get the double value of the seconds and
wrong time and date (could be read from sysfs).
To fix that the patch adds checking whether
the real_seconds was alredy set with rtc_time_sec.

Change-Id: I913421e7377732a0648a69ceadaac6e765c2613e
Signed-off-by: Oleg Kosheliev <oleg.kosheliev@globallogic.com>