drm: rcar-du: Fix display starting procedure
Because the initialization order of vsp and display are changed,
there is problem that frame end interrupt of vsp immediately
is not detected. So, before starting the display with the DEN bit,
set the DRES bit and modify the register so that it is always
reflected, and this problem is cleared.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>