drm/xen-front: Set DMA ops to default, not dummy
The device is not spawn from a device tree, so arch_setup_dma_ops
is not called, thus leaving the device with dummy DMA ops.
This makes the device return error on PRIME buffer import, which
is not correct: to fix this call of_dma_configure() with a NULL
node to set default DMA ops.

Signed-off-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>