Allow to set frame rate if not streaming
Checking for vb2_is_busy returns true if the stream is stopped,
but buffers are still allocated, e.g. dequeued but not yet
freed. This prevents from frame rate change: it is a valid use-case
when user may want to start the stream again with the same buffers,
but different frame rate.

Signed-off-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>